made readme, fixed timeslot creating to only delete yours, fixed timeslot editing only showing yours, fixed password2 matching in register, cleaned up cal invitations to be a table

+++ oneonone_api/OneOnOne/TimeSlots/views.py
@@ -35,8 +35,8 @@ class TimeSlotListCreateAPIView(generics.ListCreateAPIView):
         day_id = self.kwargs.get("day_id")
         day = get_object_or_404(Day, id=day_id)
 
-        # Delete all old timeslots related to the day before creating a new one
-        TimeSlot.objects.filter(day=day).delete()
+        # Delete only the timeslots related to the day and owned by the current user
+        TimeSlot.objects.filter(day=day, owner=self.request.user).delete()
 
         # Check if the user is a participant of the calendar
         user = self.request.user
